Move cli command implementations in the cmd package
